rename testnet -> sidechain
